This includes an example that connects via the connector (clients/connect.cpp) and some bugs and other changes I decided to make while writing it.
Please note that the address it connects to is the localhost, so you
may need to either modify that or the port the connector comes up
as. I didn't test asking it to connect to itself (though I suppose
this theoretically would work...), but that's probably not what you
want to do.