THIS CODE IS COMPLETELY UNTESTED. USE AT YOUR PERIL. ONCE TESTED IT WILL BE MERGED TO MASTER
I made a iobuf like class but specialized for only read or write
queues, which simplifies the API and hopefully will get rid of a
common source of errors. Also, they use the mapped_buffer under the
hood. I only put this in the bitcoin handlers so far. If it passes
testing I'll move it to the command handlers.

As I move the rest of the code to use this (or actually a smarter
version that switches to it for larger allocations only), most copies
of data in the system will be COW, which should make things a good
deal faster.

The small fast allocator will be mempooled with COW for things below a
certain size, then switching to the mmap_buffer afterwards. The
write_queue now does smartish things to deallocate as it goes, so
templating that to do the write thing with the smarter structures is
relatively trivial.