A perl script to convert a verbatim log to sql. Right now this writes out to a sqlite file. It was done in perl so this could be trivially switched to a different database handle without changing any of the processing code.
See the top-level comment in the code for some notes on this. This
should make it easier to answer at least some questions about the data
as it comes in.

Also, note that some views are created to make working with it
sql-wise a bit easier.

This is NOT idempotent. Rerunning it on a log will result in duplicate
records. See comment in file for rationale.