one line fix to the cycle bug I saw. It wasn't in the connector
I also put in a line in the connector to warn if something that
looks like this bug is happening. I noticed that the get_nodes
utility was outputing the handle_id to the console in the wrong
endianness. This isn't a problem because it wrote it out correctly
otherwise, but should be fixed for demo purposes