I added an extra field to all of the log messages.
This is a backward compatibility breaking change, but in the future if
you use the log_format struct it'll unpack things a bit for
you. Anyway, I updated the python library and the bcwatch utility. If
you were/are using those then if there are no bugs you wouldn't be
affected.

The purpose of this is to distinguish between messages between more
than one connector (for instance, handle_id is unique per connector,
not unique for all connectors).

My local bitcoin instance did not return any addrs in response to
getaddr, so the getaddr utility us untested with these new changes. If
you test this and all is well before I merge to master/static let me
know. It's the only reason I haven't merged yet.