I just took the obvious shortcuts with respect to performance.
Now this script runs in about 65% of the runtime of the previous
iteration and it gives you a better idea of progress as it runs.

Note: instead of taking STDIN it now takes a filename. The stream
makes more sense if it is feeding off the log directly, but thus
far that doesn't seem to be how we are doing things, and this way
I can give a progress indicator. Changing back to STDIN is trivial.