This commit contains the code for getaddr_wrapped.
In running that it became clear that there were some performance
issues with either how it was ran or in the connector itself. Both
were changes to reflect that.

The connector now does more IO buffering and getaddr_wrapped throttles
its new connections a bit. Having 10^6 connection attempts and however
many handshakes and then getaddring on top seemed to overwhelm it.

Hopefully this fixes the issue, but I think there is a performance
cliff after which we have to do multiprocessing approaches. Previous
commit messages before the rebase follow:

commiting getaddr_wrapped and friends to test outside auguste

reduce syscalls

create file if not exists, do not truncate if it does

write half the number of write syscalls to the logger

more buffering in logging

larger preallocation, saves about 5% or so I think

performance and log truncate