Fix a bug in getaddr_wrapped.cpp where conditional was swaped.
Add PEER_RESET to reconnect condition set.

Modified bitcoin_handler to use unique_ptr for the handlers. This
codifies the intent that a handler can only exist in one place at a
time, generally either the active or inactive set. This did not fix a
heap corruption bug, but made it more clear that there are cases where
a handler commits suicide twice.

This is unexpected and probably results from how libev is handling the
queuing. Even though an item is removed from the libev queue upon
suicide, perhaps certain events are propagated anyway or are already
in some kind of event queue? I'm not sure, but setting it to
unique_ptr made it more clear that this seems to be happening. The fix
is to simply return if we receive a callback on an expired handler and
to verify that a handler is in the active set during suicide,
otherwise he is already dead.