The new getaddr-slice-O-matic
This thing will slice, dice, and add extra spice (bugs).

Anyway, it is a work in progress. It combines cycle, getaddr, getaddr_wrapped,
connection_harvester, reconnector, dedupe, and cron all in one getaddr to
rule them all.

It does this in a Shadow friendly way by getting rid of all threads and other
chicanery, to instead rely solely on libev for its multitasking ways. This
means it is more complicated than any normally programmed multi-headed
beast, but hopefully will make Shadow work easier. It remains to be seen if
this is too much in a single process or if some of these tasks ought to be
spun off in a multiprocessor style thing to still play nice in Shadow. My
guess is no, but we'll see.

I sat down and wrote most of this and fixed compilation errors that the changes
were causing in the old utilities. It doesn't compile yet, but that is expected.
My intention was to first just write the thing to do and set it up to
intentionally fail the compile if it tried to do copies of structres that I do
not want to copy, so I can fix that crap later. Anyway, I don't think the logic
is complete (the cron feature is missing entirely, for instance), but it is
reasonably close. Almost all of the pieces have been written and integration
work has been done. Finalizing integration and adding some of the scraps is what
is remaining, as well as fixing my no-copy stuff to just use the unique_ptr stuff
as planned.